backend/requirements/base.txt
django==5.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
django-extensions==3.2.3
djangorestframework-simplejwt==5.4.0
django-rest-passwordreset==1.5.0
django-filter==24.3
gunicorn==23.0.0
black==24.3.0
requests==2.32.4
ipython
ipdb
